bj10986.java

+++ src/baek/gold/bj10986.java
@@ -14,8 +14,7 @@ public class bj10986 {
         int N = Integer.parseInt(st.nextToken());
         int M = Integer.parseInt(st.nextToken());
 
-        int[] rest = new int[M];
-
+        long[] rest = new long[M];
         long sum = 0;
 
         st = new StringTokenizer(br.readLine());
